requirements.txt
python-telegram-bot
spacy
